% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize.R
\name{initialize}
\alias{initialize}
\title{Initialize Outbreak Reconstruction}
\usage{
initialize(
  n_subtrees = NULL,
  n_global = 100,
  n_local = 100,
  sample_every = 100,
  root = NA,
  N = NA,
  record = c("n", "h", "seq", "N_eff", "mu", "pi", "R"),
  filters = NULL,
  indir = "input_data",
  a_g = 5,
  lambda_g = 1,
  a_s = 5,
  lambda_s = 1,
  psi = 0.5,
  init_mu = 2e-05,
  fixed_mu = FALSE,
  init_N_eff = log(100),
  fixed_N_eff = FALSE,
  init_R = 1,
  fixed_R = FALSE,
  init_pi = 0.5,
  fixed_pi = FALSE,
  ongoing = TRUE,
  safety = NA,
  split_bottlenecks = FALSE
)
}
\arguments{
\item{n_subtrees}{How many subtrees to parallelize over. Defaults to NULL, in which case it's the number of cores available to the machine, so long as each subtree has at least 25 nodes.}

\item{n_global}{Number of global iterations. Defaults to 100.}

\item{n_local}{Number of local iterations per global iteration. Defaults to 100.}

\item{sample_every}{Number of local iterations per one sample. Must divide n_local.}

\item{root}{If not NA, the sequence name given is treated as the root of the transmission network. Defaults to NA}

\item{record}{Parameters to be recorded in the MCMC output.}

\item{filters}{Filters for within-host variation data. List consisting of three named values: af, dp, and sb, meaning minor allele frequency threshhold, read depth threshhold, and strand bias threshhold, respectively. Defaults to NULL, in which case these filters are set to 0.03, 100, and 10, respectively.}

\item{indir}{Name of the directory in which we have fasta(s), metadata table, and (optionally) vcf files. Defaults to "input_data."}

\item{a_g}{Shape parameter, generation interval. Defaults to 5.}

\item{lambda_g}{Rate parameter, generation interval. Defaults to 1.}

\item{a_s}{Shape parameter, sojourn interval. Defaults to 5.}

\item{lambda_s}{Rate parameter, sojourn interval. Defaults to 1.}

\item{psi}{Second parameter of the negative-binomially distributed offspring distribution. Defaults to 0.5.}

\item{init_mu}{Initial value of the mutation rate, in substitutions/site/day. May be fixed using fixed_mu = TRUE, or inferred otherwise. Defaults to 1e-6.}

\item{fixed_mu}{If FALSE (the default), the mutation rate is estimated. If TRUE, the mutation rate is fixed at its initial value for the duration of the algorithm.}

\item{init_N_eff}{The growth rate of the within-host effective population size. Specifically, at time t after inoculation, a host has exp(N_eff * t) virions in their body. Defaults to log(100).}

\item{fixed_N_eff}{If FALSE (the default), the within-host population growth rate is estimated. If TRUE, the within-host population growth rate is fixed at its initial value for the duration of the algorithm.}

\item{init_R}{Initial value of the reproductive number. Defaults to 1.}

\item{fixed_R}{If FALSE (the default), the reproductive number is estimated. If TRUE, the reproductive number is fixed at its initial value for the duration of the algorithm.}

\item{init_pi}{Initial value of the sampling rate. Defaults to 0.5.}

\item{fixed_pi}{If FALSE (the default), the sampling rate is estimated. If TRUE, the sampling rate is fixed at its initial value for the duration of the algorithm.}

\item{ongoing}{TRUE if the outbreak is ongoing, or FALSE if has it concluded (i.e. no more samples will ever be collected from this outbreak). Defaults to TRUE.}

\item{safety}{Either NA or a non-negative value. If NA, safety mode (checking that the likelihood is correct after each global iteration) is disabled. If numeric, the maximum tolerance for a difference in the re-computed versus stored likelihood before throwing an error. Defaults to NA.}

\item{split_bottlenecks}{If TRUE, likelihood accounts for split bottlenecks (new beta feature). Defaults to FALSE.}
}
\value{
The initial configuration of the Markov Chain.
}
\description{
This function initializes the outbreak reconstruction algo.
}
