% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_mcmc.R
\name{local_mcmc}
\alias{local_mcmc}
\title{Internal function to run local moves on a subtree}
\usage{
local_mcmc(mcmc, data)
}
\arguments{
\item{mcmc}{Current state of the Markov chain.}

\item{data}{Fixed parameter values supplied to the Markov chain.}
}
\value{
List of samples from the local tree.
}
\description{
Internal function to run local moves on a subtree
}
