% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{run_mcmc}
\alias{run_mcmc}
\title{Run Outbreak Reconstruction}
\usage{
run_mcmc(init, noisy = F, logging = F)
}
\arguments{
\item{init}{Initial config generated by initialize().}

\item{noisy}{Whether to plot the current state of the Markov Chain at each iteration, as well as print the current log-likelihood, substitution rate, and mutation rate.}
}
\value{
The MCMC output.
}
\description{
This function runs the outbreak reconstruction algo.
}
