% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize.R
\name{initialize}
\alias{initialize}
\title{Initialize Outbreak Reconstruction}
\usage{
initialize(
  n_subtrees = NULL,
  n_global = 100,
  n_local = 100,
  sample_every = 100,
  init_mst = FALSE,
  init_ancestry = FALSE,
  record = c("n", "h", "w", "t", "b", "a_g", "lambda_g", "a_s", "lambda_s", "mu", "p",
    "v", "lambda", "rho", "psi"),
  filters = NULL,
  check_names = TRUE,
  virus = "SARS-CoV-2",
  a_g = 5,
  lambda_g = 1,
  a_s = 5,
  lambda_s = 1,
  rho = Inf,
  R = 1,
  growth = NULL
)
}
\arguments{
\item{n_subtrees}{How many subtrees to parallelize over. Defaults to NULL, in which case it's the number of cores available to the machine, so long as each subtree has at least 25 nodes.}

\item{n_global}{Number of global iterations.}

\item{n_local}{Number of local iterations per global iteration.}

\item{sample_every}{Number of local iterations per one sample. Must divide n_local.}

\item{init_mst}{Should we initialize to a minimum spanning tree? (Set to FALSE for large datasets due to long runtime.)}

\item{init_ancestry}{If TRUE, the initial ancestry is specified in a one-column .csv file called ancestry.csv, where the entry in the ith row is the initial ancestor of host i.}

\item{record}{Parameters to be recorded in the MCMC output.}

\item{filters}{Filters for within-host variation data. List consisting of three named values: af, dp, and sb, meaning minor allele frequency threshhold, read depth threshhold, and strand bias threshhold, respectively. Defaults to NULL, in which case these filters are set to 0.03, 100, and 10, respectively.}

\item{check_names}{If TRUE, checks whether all of the names in the FASTA match the names of the VCFs and dates.}

\item{a_g}{Shape parameter, generation interval. Defaults to 5.}

\item{lambda_g}{Rate parameter, generation interval. Defaults to 1.}

\item{a_s}{Shape parameter, sojourn interval. Defaults to 5.}

\item{lambda_s}{Rate parameter, sojourn interval. Defaults to 1.}

\item{rho}{Overdispersion parameter. Defaults to Inf, indicating the offspring distribution is Poisson.}

\item{R}{Reproductive number (average over entire outbreak). Defaults to 1. Exactly one of R and growth may be specified; the other must be set to NULL.}

\item{growth}{Exponential growth rate of cases. Defaults to NULL. Exactly one of R and growth may be specified; the other must be set to NULL.}

\item{name}{description}

\item{Virus}{being studied (affects various presets). So far, options are "SARS-CoV-2" and "H5N1." More to come soon!}
}
\value{
The initial configuration of the Markov Chain.
}
\description{
This function initializes the outbreak reconstruction algo.
}
