% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{summarize}
\alias{summarize}
\title{Summarize Outbreak Reconstruction}
\usage{
summarize(results, burnin = 0.2)
}
\arguments{
\item{results}{List returned by run_mcmc().}

\item{burnin}{Proportion of MCMC iterations to discard as burnin. Defaults to 0.2.}
}
\value{
A list consisting of a matrix of direct transmissions and their posterior probabilities, a matrix of indirect transmissions and their posterior probabilities, and posterior samples of mu (mutation rate in substitutions per site per day), p (mutation rate in mutations per site per cycle), b (probability of incomplete bottleneck), and the time of the MRCA (if the tree is unrooted).
}
\description{
This function summarizes the MCMC output.
}
